Fix windows

+++ ipykernel/kernelbase.py
@@ -16,8 +16,13 @@ import uuid
 import warnings
 from datetime import datetime
 from functools import partial
-from signal import (SIGINT, SIGKILL, SIGTERM, Signals, default_int_handler,
-                    signal)
+from signal import SIGINT, SIGTERM, Signals, default_int_handler, signal
+
+if sys.platform != "win32":
+    from signal import SIGKILL
+else:
+    SIGKILL = None
+
 
 try:
     import psutil
@@ -1149,9 +1154,10 @@ class Kernel(SingletonConfigurable):
                 await asyncio.sleep(0.05)
                 self.log.debug("Sending SIGTERM to {pgid=}")
                 os.killpg(pgid, SIGTERM)
-                await asyncio.sleep(0.05)
-                self.log.debug("Sending SIGKILL to {pgid=}")
-                os.killpg(pgid, SIGKILL)
+                if sys.platform != "win32":
+                    await asyncio.sleep(0.05)
+                    self.log.debug("Sending SIGKILL to {pgid=}")
+                    os.killpg(pgid, SIGKILL)
             except Exception:
                 self.log.exception("Exception during subprocesses termination")
             return
@@ -1163,7 +1169,12 @@ class Kernel(SingletonConfigurable):
             return
         self.log.debug(f"Trying to interrupt then kill subprocesses : {children=}")
         self._send_interupt_children()
-        for signum in (SIGTERM, SIGKILL):
+        if sys.platform != "win32":
+            sigs = (SIGTERM, SIGKILL)
+        else:
+            sigs = SIGTERM
+
+        for signum in sigs:
             self.log.debug(
                 f"Will try to send {signum} ({Signals(signum)}) to subprocesses :{children}"
             )
